feat(deepseek): add --cast_mxfp4_to_nvfp4 to deepseek_v4 quantize step
Add a closed-form, bit-exact MXFP4 -> NVFP4 routed-expert weight cast to examples/deepseek/deepseek_v4/quantize_to_nvfp4.py via a --cast_mxfp4_to_nvfp4 flag. Pins scale_2 = 2^(k_max-8) and each per-block E4M3 scale to 2^(k_j-m) from the source E8M0 scales, so the NVFP4 nibbles equal the source MXFP4 nibbles bit-for-bit for every in-range block. w1/w3 share one scale_2 for the fused GEMM1; activation input_scale still comes from --amax_path calibration.

Hoist the shared closed-form numerics (mxfp4_to_nvfp4_global_amax, mxfp4_to_nvfp4_per_block_amax, and the E2M1/E4M3/E8M0 constants) out of the GPT-OSS example cast (examples/llm_ptq/cast_mxfp4_to_nvfp4.py, PR #1372) into modelopt.torch.quantization.utils.numeric_utils, so both the GPT-OSS and DeepSeek-V4 cast paths import them from the library. Their unit tests move to tests/unit/torch/quantization/test_numeric_utils.py; the example test keeps the cast-specific cases (quantizer naming, build_amax_map, apply_to_model).
